requirements.txt
fastapi
python-multipart
google-generativeai
